Update Width Correction when section_num == 1
When there is only one section, we do not need to retrieve the machine width from the field_data. The field data doesn't specify units for the machinery width, but the current code assumes the machinery width is in feet. I think it is better to grab the width from the asapplied data, where we have already identified the units and converted the width to meters. This also avoids allowing any entry error in field_parameters to affect the results.