Adding lines to not lose farmer prices when running unpack_field_parameters
When we bring the field_parameter.json into this code, sometimes the crop_price changes from numeric in the json to string in R. This results in the default price being used instead of the farmer's price. I'm adding as.numeric() to the process of defining the prices. This will force any strings, such as those from the default field template, to NA, which will receive the default price.